pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vertexai-search-media"
version = "0.1.0"
description = "Vertex AI Search for Media - Specialized media search and recommendation CLI"
authors = [{name = "Drama Shorts Startup"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "google-cloud-discoveryengine>=0.11.0",
    "google-cloud-storage>=2.10.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "jsonschema>=4.19.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "google-cloud-bigquery>=3.0.0",
    "python-dateutil>=2.8.0"
]

[project.scripts]
media-data-store = "src.media_data_store.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
